release jobs where start.pl died early on
start.pl might die early due to some syntax error or misconfiguration.
Take existance of results.json as indicator of successful start. If it's
missing when start.pl died, release the job instead of setting it to a
final state. When doing that worse the priority value so it's not
grabbed right away before any other job. If the prio is above a
threshold, give up and set it to failed.