action #3582 - openqa: make dents work again
This patch show the dents in the user interface.  After talking
with coolo and lnussel we agreed that this is not the best approach:

os-autoins needs to mark as fails the test when there is a dent.

This patch is submitted meanwhile, because can be used to track the
dents.