limit_assets: check asset dirs exist before accessing them
We shouldn't just blindly die if we can't open a directory
that doesn't even exist, just in order to check if some files
in it are registered assets. If it doesn't exist, it's not going
to contain non-registered assets.

This is perhaps a bit academic for /iso, which is probably
always going to exist, but it isn't for /repo, which is SUSE-
specific at present and entirely optional.