allow ISO to be a URL, download if needed
This adds an optional 'ISOURL' parameter for the ISO POST API
call. If ISOURL is passed, the jobs will be created, and a gru
task to download the ISO will be scheduled. A dependency
between the two is registered in a new GruDependencies table.
When the gru task completes, the GruDependencies entry will be
deleted. job_grab() excludes jobs whose IDs are listed in the
GruDependencies table.