Improve web UI branding mechanism, drop topbar in 'plain'
This introduces an improved mechanism for branding-specific
elements in the web UI templates. Instead of just peppering
'if' blocks into the templates, this introduces a helper
which works like 'include', only it understands branding.
It accepts a name for the template to be included. It reads
the 'branding' config setting. It then tries to load the
template 'branding/(brandname)/(templatename)'. If that
doesn't exist, it falls back to loading 'branding/plain/
(templatename)' - e.g. if the 'branding' setting is
'openSUSE' and the call is 'include_branding topbar', it
will try to load 'branding/openSUSE/topbar' first; if that
doesn't exist, it will load 'branding/plain/topbar'.

The existing 'if openSUSE then (foo), else (bar)' blocks
are replaced with calls to the new helper, and the actual
contents are placed in the appropriate locations.

This means anyone can create a new branding by just adding
a subdirectory with appropriately named files and changing
the config setting. It also means such custom brandings
can be partial, because of the fallback to 'plain' - if you
just want to stick one branded element in on top of plain,
you can easily do that.

As part of the change, the openSUSE topbar that was present
on all pages (with geeko and the 'Downloads', 'Support' etc.
links) even with the 'plain' branding before is now only
present in the openSUSE branding.