set 0640 perms on sqlite db when creating it
It seems to me that the openQA database really should not be
world-readable, since it contains the API keys (and perhaps
other sensitive stuff, I dunno). This changes initdb to set
0640 perms on the sqlite database file when creating it. Would
probably go along with a change in the spec file to set the
same perms on the %ghost line for the database file. For more
advanced databases I guess access control should be up to the
admin.