move 'latest jobs' logic from overview to ResultSet::Jobs
This is my first cut at #436. It takes the 'latest jobs' logic
from the 'overview' web UI function and makes it into a method
in ResultSet::Jobs. Then it exposes it via the web API: you
can pass 'latest' as a query parameter (set to anything) and
only the latest jobs will be included in the results. The logic
works as before.

This is broadly along the lines suggested by @coolo in #436,
but I'm not sure how happy I am with the result, improvement
suggestions welcome. We *could* still put this in query_jobs
by having it iterate through the resultset and find the IDs of
the 'latest' jobs then produce a new resultset filtered by
those IDs, but I'm not sure if that'd be better.