Add PRODUCTDIR variable to support multiple main.pm/needles
Allow to accomodate multiple "main.pm" and accompanying needles directory
depending on product while still sharing same tests.

PRODUCTDIR is discovered from trying to find directories inside "casedir".
PRODUCTDIR is forwarded to os-autoinst along with CASEDIR.

If a test distribution with the new folder structure is used it needs support
for PRODUCTDIR in os-autoinst. It should be save to install this new version
of openQA in an environment with old os-autoinst as well as an old set of
test.

Verified with a local installation of os-autoinst, openQA and
os-autoinst-distri-opensuse.

Related os-autoinst change: https://github.com/os-autoinst/os-autoinst/pull/376