Add 'coverage-check' script for local coverage threshold test
Add script 'check_coverage' which parses the HTML coverage report for the
'total' coverage. The check fails with a notification message if the coverage
drops below the threshold defined in the Makefile.

This way we can prevent further unintended coverage drop while allowing to
adjust the threshold when necessary. However we should try to increase the
threshold over time which could also be done by marking "uncoverable"
statements.