JobModules details(): catch malformed JSON files (#517)
On Fedora's openQA instance, some of these JSON files are
malformed, for some reason. This causes the GRU scan_old_jobs
task to fail, and GRU tasks failing is really bad, they just
repeat on an endless loop. We already guard against not being
able to open the file and return an empty list, so it seems to
make sense to do the same if the file is malformed.