job_schedule_iso: fix when ISOURL is set but not ISO
so, er, oops. this has been broken since initial commit. I
always intended it to be possible to post a request with just
ISOURL set, but not ISO, if you just wanted openQA to download
the ISO with its original filename; setting ISO along with
ISOURL is supposed to be optional and only necessary if you
want to download the ISO with a *different* filename. However,
I must have wound up only testing the case where both are set.
If only ISOURL was set, then two things wound up broken:

* the download target location was incorrect as the code to
set it assumes $args{ISO} is specified, so the download
subroutine would never download the image because the target
location already exists (it winds up being just `isodir`)

* the generated jobs don't have an ISO setting, so even if you
fixed the first problem, the downloaded ISO wouldn't actually
be attached to the VM

this fixes both problems and makes things actually work, when
a request is POSTed with just ISOURL set.