log to stderr (so journal) by default
in Fedora package review it became apparent that logging to a
file is rather more complex in terms of packaging and system
security than it might appear. Logging to the journal is kinda
a good practice on systemd distros (i.e. all the ones we care
about) in any case, and we can easily default to it by leaving
the 'file' setting undefined by default. Admins who want to log
to a file still easily can by defining the 'file' setting, and
then the security issues are all theirs to deal with...

This isn't very sophisticated (the severity of the messages
isn't transmitted to journald, for instance) but works well
enough, seemingly.

This change will affect existing deployments which do not have
the `file` setting in the config file explicitly specified.