Add generate-packed-assets script for asset generation
This is to support Zbigniew's proposal to do asset generation
with RPM triggers instead of in %install - see:
https://bugzilla.redhat.com/show_bug.cgi?id=1304882#c30
If we think that's a good idea, it makes sense to have the
script in openQA rather than adding it to packages as a separate
SOURCE. Even if someone doesn't want to use triggers or their
distro's equivalent, they can call this script instead of having
the logic in the package build, so at least we're all using the
same code.

I can't personally see any security implications in allowing
specification of the prefix as an arg like this, but please yell
if you can. I *think* anyone who could manipulate the call to
the script to call some other executable would have much easier
ways to exec stuff anyway - they'd have to be able to modify
the package or the script itself, and if they can do either of
those things, you're already boned.