docs: Update Contributing guide
"rules for commits" is an adaption of
https://github.com/os-autoinst/os-autoinst#rules-for-commits
for openQA. Initially I thought we should reference
"os-autoinst" rules but as the tidy path as different as
well as the Makefile and the tests take longer than
os-autoinst I adapted the section for openQA and included
it as another subsection splitting the current
"Development" guidelines in a general section and two
subsections "Rules for commits" and
"Getting involved into development".

Old references to two-week sprints and google hangout
calls are removed and replaced by a reference to the IRC
channel and a notice about a "jangouts" call for the
core developer team without further details.

One reference to "openQAv2" which was pointing to
"openqav3" on progress.opensuse.org is corrected to
"openQAv3".