Show bug or label icon on overview if labeled
As we parse the comments for generating the overview page we can now
do some simple parsing and show a "bug" icon if a corresponding reference in
a comment has been found or a "label" icon if the keyword "label:" has been
used. This way we can show bug references using the convention
"<bugtracker_abbreviation>#<bug_id>", e.g. "bsc#1234" or labels in the
(not yet) convention "label:<single_word_label>", e.g. "label:false_positive"
or "label:broken_test".

Priority is as follows: bug > label > any comment. Based on this only one icon
is shown not to clutter up the overview page.

Currently the behaviour based on multiple bugs or labels found is undefined.