Show bug icon with URL if mentioned in test comments
Extract util methods for resolving common bug references to a complete bugurl
or an HTML 'a href' tag. Add a helper 'bugurl_for' for use in templates and
use this helper within 'test/overview.html.ep' to make the bug icon a link.

The use of "OpenQA::Utils" within "OpenQA::WebAPI::Plugin::Helpers" - which
does not look like a good "plugin" candidate itself - might be disputable or
a candidate for refactoring.