tests: don't explode if Selenium isn't there
The phantomjs tests are supposed to be kinda optional - they're
skipped gracefully if phantomjs isn't there. However, there's a
couple of dependencies on Selenium bits which aren't similarly
gracefully handled; even if the system doesn't have phantomjs
installed so the tests would wind up skipped anyway, the suite
winds up blowing up if Selenium::Remote::Driver and Selenium::
Remote::WDKeys aren't present.

This should fix that. It uses Module::Load::Conditional to see
if the modules are available; it's a dependency of IPC::Cmd, so
we're not really adding any extra dep here.

We also adjust the messages to explain that Selenium bits are
needed as well as phantomjs, and add IPC::Cmd to the reqs in
cpanfile.