systemd: start scheduler (and webui and websockets) after dbs
If the database is running on the same system, we want to make
sure it's up before we try and start the scheduler, webUI and
websockets. As we support both postgre and mariadb, list both.
gru already has postgresql, but add mariadb for it.

webui and websockets currently are ordered after scheduler, so
we don't really need to do this in all three, but we may as
well just in case that changes.