Show certificate next to builds on overview if all failures are labeled
Based on comments in the individual job results for each build a certificate
icon is shown on the group overview page as well as the index page to indicate
that every failure has been reviewed, e.g. a bug reference or a test issue
reason is stated. Only the failed and incomplete jobs are regarded for the
evaluation if a build is considered "reviewed".

Tested against a production database dump and no performance impact could be
found when rendering the index page.