Do not carry over labels over passes
After a job passed a new issue in a subsequent fail is assumed to be failed
for a different reason.

The test is restructured into subtests for the individual scenarios.

For restarting the jobs in tests the test detail templates showed an error
because of undefined job start time so we just set it to zero duration in
these cases.