Carry over labels from previous jobs in same scenario if still failing
It is possible to label all failing tests but tedious to do by a human user
as many failures are just having the same issue until it gets fixed.
It helps if a label is preserved for a build that is still failing. This
idea is inspired by
https://wiki.jenkins-ci.org/display/JENKINS/Claim+plugin

The scheduler might not be a good location for "_carry_over_labels" so this
method could also be moved elsewhere.

Related issue: https://progress.opensuse.org/issues/10212