Init or upgrade schema from openQA
Preparing the database upgrade is the same as before, but manual
initdb resp. upgradedb call to perform initialization resp. upgrade
is not needed.
Upon first connect, openQA will check if database contain schema version.
If none is found, db is initialized. Else it is compared to current available
schema version and if not matching, upgrade is done automatically.