Add optional query parameters for showing previous builds
The index dashboard and group overview page limit the number of previous
builds based on hardcoded numbers. Adding query parameters this is easily
adjustable e.g. if one wants to get an overview of a longer history.

Example to show up to three week old builds instead of the default two weeks
with up to 20 builds instead of up to 10 being the default for the group
overview page:

    http://openqa/group_overview/1?time_limit_days=21&limit_builds=20