Optimize comment lookup for group_result by prefetch
Same functionality as in before but faster as comments are prefetched for all
jobs in each build in each group. Obviously this could be further optimized
by prefetching comments for all jobs at once but we do not know the jobs
before iterating the builds.

Tested with a production database dump from openqa.suse.de from
2016-02-25 fetching the page 'group_overview/25'. Comparing the number of
queries for comments and the user space time.

Before: 356 @ 2.6s
After: 11 @ 1.9s