Optimize overview page rendering by only acting on latest jobs
The 'latest_jobs' are queried anyway and gathering job module stats as well
as calculating the preferred machines should be done on latest jobs. Acting
on all potential jobs returned by the query is a stupid idea especially if
there are many jobs within a build but not necessarily the latest.

This fixes a performance regression when calling the overview page on
openqa.suse.de for some builds.

Performance was measured retrieving
'/tests/overview?distri=sle&version=12-SP2&build=1204&groupid=25'
with a database dump from openqa.suse.de from 2016-02-26.

Performance results:

    before: 120s
    after: 27s

Obviously "27s" is still not what we would desire or what we are used to in
other builds but this is the "worst case scenario" I found.

The four most expensive queries are:

before:
12.45848107 seconds executed: SELECT ... FROM job_modules
5.55346823 seconds executed: SELECT ... FROM job_settings
5.53332710 seconds executed: SELECT ... FROM job_settings
4.64096379 seconds executed: SELECT ... FROM job_settings

after:
4.69238210 seconds executed: SELECT ... FROM job_settings
0.00488114 seconds executed: SELECT ... FROM job_modules
0.00128818 seconds executed: SELECT ... FROM comments
0.00120902 seconds executed: SELECT ... FROM job_modules

Related issue:
https://progress.opensuse.org/issues/10966