Reintegrate the ISO scheduling into the ISO controller
Relying on DBUS for this is actually more fragile - and the reason we
did it this way was to avoid locks and such. But if the WebAPI creates
the jobs, it can actually run the whole thing in a transaction, so if
we fail, we can rollback. With the DBUS service scheduling, the WebAPI
can never be sure