Tag builds with special comments on group overview
Based on comments on the group overview individual builds can be tagged. As
'build' by themselves do not own any data the job group is used to store this
information. A tag has a 'build' to link it to a build. It also has a 'type'
and an optional 'description'. The type can later on be used to distinguish
tag types.

A 'tag' icon is shown next to tagged builds together with the description on
the group_overview page. The index page is not changed to prevent a potential
performance regression.

Within the sub group_overview the comments are parsed for comments and then
passed to the template explicitly to prevent duplicate database queries.

Related progress issue: https://progress.opensuse.org/issues/9544