Do not clean up logs of important builds in GRU task
As builds can now be tagged we come up with the convention that the
'important' type - the only one for now - is used to tag every job that
corresponds to a build as 'important' and keep the logs for these jobs so that
we can always refer to the attached data, e.g. for milestone builds, final
releases, jobs for which long-lasting bug reports exist, etc.

As these jobs are not cleaned up automatically a manual or external cleanup
scheme has to be applied for important builds and jobs.

Also add tests for 'reduce_result' GRU task.

Performance evaluation of getting a tests/overview page, index page and a group
overview page including the commit adding build tags.

* before: 1.5+-0.1, 2.4+-0.1, 2.2+-0.0
* after: 1.5+-0.1, 2.0+-0.1, 2.2+-0.1

Related progress issue: https://progress.opensuse.org/issues/9544