Mark the worker as alive when it uploads things
Otherwise we will kill it after 10 seconds of no keep alive signal. As the
uploading keeps the mojo IO loop busy, it's unable to send the signal reliable
and we get our legs kicked. If 10 seconds is not enough to upload (huge disks
possibly?), we need to warn openQA before we do that and store it in the
jobs or the workers table. For now this might be enough though