limit_assets: never remove assets for scheduled/running jobs
I'm not at all sure this is correct, but I *think* it may be
needed. AFAICS, limit_assets does not do anything at present
to avoid removing assets associated with pending jobs, and I
think it should, so it doesn't 'clean up' a generated HDD
while a job is still queued up expecting to use it, for e.g.

Please do have someone familiar with this code check this over
and make sure a) it's needed and b) it works right.