allow loading of arbitrary plugins from config file
For now this only loads plugins for openqa-webui, and expects
them to be placed in WebAPI/Plugin . You can specify the config
setting 'plugins' as a space-separated list of module names of
plugins to be loaded, e.g. 'plugins = Fedmsg' would cause the
plugin WebAPI::Plugin::Fedmsg to be loaded.

We might want to define a different location for external
plugins, or allow loading of plugins for other bits of openQA,
or allow some kind of way to specify arguments to be passed to
the plugin...this is just a simple starting point for a PR.

The motivation here is to let me load a fedmsg plugin for
Fedora's openQA; I could just send a PR with a special config
key like the one for the AuditLog plugin, but that doesn't
seem like a particularly extensible approach, and it'd be odd
for openQA to have code specifically for loading a plugin it
doesn't actually include (I don't think the plugin itself needs
to be part of the openQA codebase, I was expecting to maintain
it externally).