add fedmsg emitter plugin, bit more commenting for plugin load
As @aaannz is interested in using this plugin, we may as well
make it part of upstream openQA. It simply relays some openQA
signals as fedmsg signals, for now using the 'fedmsg-logger'
command as we don't have Perl bindings for fedmsg and I don't
feel a lot like writing them. You need to have a correct fedmsg
relay config in place to use this; for local testing it will
suffice to install fedmsg-relay with its default config and run
it, that will emit the messages on the local fedmsg instance.
For production use, you need an appropriate relay_inbound
setting which points to an active->passive relay for the real
fedmsg instance, and probably you'll want to ensure a cert and
key are in place for message signing, with prefix 'openqa'.

The logic for adding the 'remaining' count could perhaps move
to the place where the openQA job_done message is constructed,
I guess.