limit_assets: when sizelimit is exceeded, remove more
We (Fedora) still seem to hit edge cases with this. Sometimes
we had a lot of jobs running simultaneously and one uploaded
hdd image went missing, even though we've reduced the number of
tests that upload images as far as possible and upped the asset
size limit to 300GB. I think this may help deal with it. This
makes limit_assets reduce total asset size to 80% of the limit
when the limit is exceeded, instead of just removing enough to
get back under the limit (which, if you think about it, means
any long-lived openQA instance is perpetually teetering on the
brink of exceeding its asset size limit). This way there's a
bit of breathing room each time the size limit is hit, so it
doesn't immediately get hit *again* the next time any ISO is
downloaded or hdd image uploaded.