add a config setting to specify asset types not to link to
We (Fedora) really don't want .iso files to be linked in the
Logs & Assets tab. Fedora's openQA is not a download server and
there is a more appropriate place to download those ISOs from.
So this changes things a bit so the admin can specify a list of
types not to include links to, rather than just hardcoding it.