Test comment editing in test results, too
The test should now meet the acceptance criteria
mentioned in the progress item. However, it
would be nice to test the behavior of the commenting
feature when logged in as regular user, too.
Hence there are still some TODOs in the code.
Unfortunately I currently don't know how to
test as another user.