Replace 'warn' by log_warning to catch it properly in tests
A 'warn' does not really help here as the corresponding code is a check on
settings provided by the user. This way the output is also properly catched
during test execution and does not spill on the console. The warning is still
visible e.g. if 'prove -v' is used for executing the test.