clone_job.pl: Skip downloading assets generated by parent jobs
In case all dependencies on a job are cloned including a parent job which will
generate an asset which is used by the cloned child there is no need to
download the corresponding asset when it will get overwritten by the new
scheduled parent job anyway.

This might also help to prevent confusion when the option
'--skip-chained-deps' is not used as the qcow image is downloaded but not
used.