Distinguish product and test issues bugref
Progress is used to track test issues, bugzilla for product issues, at least
for SUSE/openSUSE. This might be a lot different for other groups so a more
flexible approach could be thought of but probably it is not as easy as
just distinguishing on the issue tracker URL being used, e.g. if a group
uses just one issue tracker but different components or issue types to
differentiate, it is not that easy. That is why this simple solution is
done here, first. At least others can follow the approach and have a test
and CSS rule available ;-)