Show scenario within results details view
The scenario is becoming more and more important in referencing results. When
working with many test results at the same time or using deep links to
failed modules it was not possible to find out the scenario from the details
view page but one had to click and load the results page.

Fixes https://progress.opensuse.org/issues/12084