Spare obsoletion of scheduled jobs with comments to keep them as 'important'
This will keep scheduled jobs that have any comment in the schedule and not
cancel them when a new build is triggered.

This can be used to mark jobs which should not be obsoleted, e.g. when you
want to call many jobs in one scenario to reproduce sporadic issues.

This actually allows to "finish important builds" but the challenge is to
add a comment to all jobs of a build which is left as an exercise to a later
commit or external script ;-)

Related issue: https://progress.opensuse.org/issues/9760