Always finish tagged aka. 'important' builds
All tests in a build that has been tagged are not obsoleted if a new build
is triggered. Actually any tag is regarded as 'important' but if it has been
tagged there is not much lost if we finish this build. The obsoletion feature
is especially important for fast paced re-scheduling when no one cares to
put a comment anywhere.

As a tag comment can be created at any time regardless if the corresponding
build exists or not builds to come can be marked to be important already
upfront.

Related progress issue: https://progress.opensuse.org/issues/11072