Make fail/softfail thumbnails more eye-catching
For fail and softfail thumbnails the border style is additionally set so that
these important steps are even more distinguishable. This way the status can
be determined without relying on color. This change has been recommended by a
person with color vision deficiency. For softfail thumbnails the border is
dotted, for fail thumbnails the border is dashed to resemble the higher
importance of fail over softfail.