[WIP] who can get this to work? -- Add static scenario route
Should always refer to most recent job for the specified scenario.

* have the same link for test development, i.e. if one retriggers tests, the
person has to always update the URL. If there would be a static URL even the
browser can be instructed to reload the page automatically

* for linking to the always current execution of the last job within one
scenario, e.g. to respond faster to the standard question in bug reports "does
this bug still happen?"