Render links on 'latest' route directly instead of redirect
This allows a '/latest' link also used as a permalink to the always latest job, especially for local machines. http://<openqa>/tests/latest w/o query parameters already gave just the job with highest id but as a temporary redirect, i.e. reloading the page does not update the job. This way for example the browser can be configured to always reload the same page to have the always recent job displayed for local test development.

To still provide a link to the specific job the infobox now has a link to the '/tests/[:jobid:]' route. We could also put this link somewhere else, however, if it is preferred.