Add web UI controls to select more builds in group_overview
The query parameter 'limit_builds' allows to show more than the default 10
builds on demand. Just like we have for configuring previous results, the
current commit adds web UI selections to reload the same page with
higher number of builds on demand. For this, the limit of days is increased
to show more builds but still limited by the selected number.