Disable auto-duplication of jobs (poo#13042)
Auto-duplication of jobs just hides errors which we want to learn and improve
on. A worker might be right to retry when any web API call fails temporarily
so we might want to add back some auto-duplication unless we can handle a
retry even at a lower level to not discard the whole job which is another
story.