Provide 'API help' menu entry
The menu entry shown for every operator links to '/routes' which shows
the standard 'debug' page as shown for '404'. This page is only available to
operators as it might leak server config in the log file snippet rendered on
the page.

The intention of providing the list of available routes is to have at least
some kind of help for users of the API without needing to ask them to read the
source code and understand the Mojolicious routing algorithms.

Related to https://github.com/os-autoinst/openQA/pull/831

Related progress issue: https://progress.opensuse.org/issues/13192