Go back to using non-blocking upload so status updates happen
Following on from a50e86a, this completes the revert of f2547e9b.
There is a reason to use the non-blocking post and then 'block'
ourselves while we wait for it to complete: this allows Mojo
timers to keep triggering while the post runs. Which means the
worker will continue to send status updates to the server, and
the server's 'dead worker' check won't decide the worker is dead.