when checksumming uploaded assets, go 100MiB at a time
just piping the entire file to a 'cksum' system call can block
the worker process for over a minute if the file is large,
again rendering us vulnerable to the 'dead worker' check, as
the worker will not send status updates to the server while
it's blocked. Instead, use Digest::MD5 (we're using it in some
other places too) and read 100MiB at a time from the file into
the digest context, updating the status every ~5 seconds. I
tried to find a way to let the timer fire while this is going
on, but I can't; this is simple enough and works.

The actual digest calculation does not take long at all, it's
reading gigabytes of data from disk that takes time.